Add piedpiper.py script to compress CSS/JS.
This script hooks into the pipe_runner.conf file, but doesn't rely on
named pipes (which don't work happily with Virtualbox, nginx, or Apache)
and doesn't require manual regeneration.

This is mentioned in the script's comments, but bears repeating here:
this is a moderately epic security hole, and should never be seen
anywhere near production servers, because anyone who can edit the
pipe_runner.conf file can execute arbitrary code on the server.