Be a bit more lenient with the virtualenv setup.
Sometimes PyPI can be a bit fiddly, so giving it some extra time to do
its thing and retrying it if it fails seems like as good a solution as
any.

If this continues to be a problem, there may need to be a bit of the
script that tries to recover from a half-installed virtualenv by
installing any missing packages.